refactor: rename composables to hooks and add a readme to the folder
