+++ src/core/hooks/README.md
@@ -1,0 +1,5 @@
+# Composables and Hooks
+
+All Vue.js composables (from the Composition API) also referred as hooks in frameworks
+like React. Hooks are a more powerful, cleaner and, more flexible version of Mixins and
+should be the go to method to share logic across App components.
