+++ src/core/pages/README.md
@@ -11,11 +11,11 @@ Check out [`vite-plugin-pages`](https://github.com/hannoeru/vite-plugin-pages) f
 For example, instead of having
 
 ```ts
-import { isDark } from '../../../../composables'
+import { isDark } from '../../../../hooks'
 ```
 
 now, you can use
 
 ```ts
-import { isDark } from '~/composables'
+import { isDark } from '~/hooks'
 ```
